cloud_rangers/project/food_analysis/requirements.txt
streamlit
pandas
pillow
opencv-python-headless
pyzbar
openfoodfacts
google-generativeai
requests
